Revert "In progress"
This reverts commit 20797cee4582485776a50e4357bb0a9903751976.

+++ TestSutraUsers/src/main/java/com/tathvatech/site/oid/ProjectSiteConfigOID.java
@@ -5,7 +5,7 @@ import com.tathvatech.ts.core.common.OID;
 
 public class ProjectSiteConfigOID extends OID{
 
-	public ProjectSiteConfigOID(long pk, String displayText)
+	public ProjectSiteConfigOID(int pk, String displayText)
 	{
 		super(pk, EntityTypeEnum.ProjectSiteConfig, displayText);
 	}
